Fix localization implementation for missing keys
